refactor(tests): split PositionBookTests (445→241) (#272)
Move the investment-rule suite (dailyBalance investment rules +
apply(_:asStartingBalance:)) into a sibling PositionBookInvestmentRuleTests
file — those tests share the investmentAccount UUID and belong to the
investment-rule domain, forming one coherent group.

PositionBookTests.swift keeps the core mechanics (Empty/Apply, cleanZeros)
and the non-investment dailyBalance tests (single-instrument,
multi-instrument, earmark clamping).

Drops the file_length (445→241) and type_body_length (343) baseline
entries entirely; no other baselined rules were keyed to this file, so
nothing else needed fixing. No access widenings — each sibling keeps its
own private fixtures.
